// lib/generated/AccessibilityProfileOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DistrictHourlyProfile } from "./DistrictHourlyProfile";

export type AccessibilityProfileOutput = { generatedAt: string, maxMinutes: number, mode: string, districts: Array<DistrictHourlyProfile>, };