// lib/generated/DistrictHourlyProfile.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DistrictHourlyProfile = { name: string, key: string, hours: Array<number>, reachableCells: Array<number>, peakHour: number, peakCells: number, troughHour: number, troughCells: number, serviceDropPct: number, };